Try to restrict USB early in boot
Try to turn off USB access during early boot, and ensure that the HAL's
mUsbDataEnabled is updated appropriately so that the framework is aware
that it is off and is able to turn it back on when needed.

TODO: When done testing, change to use ro.debuggable instead (currently
commented out, but tested working as expected).

Issue: calyxos#1414
Co-authored-by: Tommy Webb <tommy@calyxinstitute.org>
Requires: Ic2f1debeeca1492125ab9a24d597f7d2f056d1cc
Change-Id: I192db722282fb72c59c25a4dac992cd09209f994