Wi-Fi Calling: Proprietary kitchen sink
Everything from lineage's proprietary-files.txt except:
system_ext/priv-app/ConnectivityThermalPowerManager/ConnectivityThermalPowerManager.apk

This will obviously need to be narrowed down to the bare minimum and
right now includes many things that are not even related at all.

It may be possible for us to get by with WfcActivation.apk alone, as
long as sepolicy is fixed, but this has not yet been tried.

Change-Id: I6cce0662f1adcc1f552cb9d6df80b81f123422b5